CW1 Updating SL Length Method
The songlist length option now outputs values to a decimal value as to provide a more accurate time. These decimal values are rounded to 2 decimal places as to be easier to read and understand; for example, if the songlist was an hour and a half, it would be output as 1.5 hours.